Try disabling laft Backspace
As my left hand is considerably weaker than my right hand (and twising
the palm results in discomfort - that's why I've changed the location of
V), I want to try unlearning the left backspace and see whether I can
live with that. Also, this may be an opportunity to reclaim this key for
other purposes.